# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GameTheca API",
    "version": "0.2.0",
    "description": "Self-hosted multi-user game library API. Authenticate with session cookies or Authorization: Bearer gt_<prefix>_<secret>."
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gt_token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
...
  },
  "paths": {
...
    "/api/chat/channels": {
      "get": {
        "summary": "List household channels + accessible DMs",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Channels" } }
      },
      "post": {
        "summary": "Create household channel (admin/librarian)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/chat/dm": {
      "post": {
        "summary": "Open or create a DM thread",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "DM channel" } }
      }
    },
    "/api/chat/channels/{channel_id}/mute": {
      "post": {
        "summary": "Mute or unmute a channel/DM for the current user",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "channel_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Muted state updated" },
          "400": { "description": "muted required" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/chat/channels/{channel_id}/messages": {
      "get": {
        "summary": "List messages (includes reaction counts + attachments)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "channel_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "before", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Messages" } }
      },
      "post": {
        "summary": "Post a message (optional attachment_ids from prior upload)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "channel_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/chat/channels/{channel_id}/attachments": {
      "post": {
        "summary": "Upload pending chat attachment (multipart file); child accounts 403",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "channel_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "201": { "description": "Pending attachment + limits" },
          "400": { "description": "Size/MIME/empty reject" },
          "403": { "description": "Child upload ban or no channel access" }
        }
      }
    },
    "/api/chat/messages/{message_id}/reactions": {
      "post": {
        "summary": "Toggle emoji reaction (fixed Unicode + custom :slug:)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "message_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Updated counts" } }
      }
    },
    "/api/chat/emoji": {
      "get": {
        "summary": "List fixed + household custom reaction emoji",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Emoji catalog" } }
      },
      "post": {
        "summary": "Upload custom emoji (admin)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" }, "403": { "description": "Admin only" } }
      }
    },
    "/api/chat/emoji/{slug}": {
      "delete": {
        "summary": "Delete custom emoji (admin)",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/api/chat/search": {
      "get": {
        "summary": "Search message bodies across accessible channels",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 2 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Search hits" } }
      }
    },
...
  }
}